Add E2E testing for all pushes
The previous set up would only run the Jest tests post deployment to github
pages on the live website and deployments only occur on the `main` branch.
This will run a python webserver on the Github runner and run the tests on
that. Most important, all branches will now have testing run.