Use better image formats
Added usage of `<picture>` tags to html files to select best possible
user-supported image. Got rid of usage of large png files entirly, jpg is just
as compatitable. Usage of images in the js and css is replaced with the jpg
versions.

TODO: Like the picture tag, in the JS and CSS we should use the best format
possible. At the moment they're hardcoded to use one particular image. FIXME.
See `imageUrl` in `source/common/nav.js` and the various styles.css with
`background-image` being used on a `url()`