Add avif; Use better images in fortune-cookie and palm-reader; Fix right zodiac wheel
Added avif images and the appropriate `<source>` tags. Added `<picture>` tag
usage for fortune-cookie and palm-reader (didn't get to it last time). Also it
seems I missed one of the zodiac wheels for putting it in a `<picture>` tag.
This has been fixed.

TODO: FIXME: I couldn't put the arrows in love-compatibility inside picture
tags, doing so seems to break their positioning. This isn't the biggest of
deals since those are small images anyways. Honestly they should probably just
be replaced with an svg instead (and we can rotate the same image to make both
left and right versions too).