Replace the settings icon with svg
Replace the settings icon with an svg to improve performace. Also make
it white so you can see it better while we're at it. Also moved the
location to be consistent with the other svgs.

Icon is from https://github.com/coreui/coreui-icons which is licensed
under CC0 1.0. Basically means we can add it without attribution required.