HACK: Fix failed Jest test
home-page.test.js is failing on the `checking to make sure the link actually works`
bit because page.click() works by clicking where the `a.team-link.hover-underline`
is. However, now that there are animations, it seems it's clicking while the
animations are still going. Which leads to click on the Fortune cookie card.
As a quick fix, I literally just made the tests wait 2 seconds before starting.
A real solution would wait for the animations to finish, but I don't know how to
do that.