Disable Japanese default selection
I tested the translations on mobile. Chinese simplified and traditional
are both fine, however Japanese is borked in many cases because the
text is now too long and sizes of things are using absolute units. It
is fine on desktop though so I decided to make it so the langauge
auto-detection won't set the site to Japanese if your browser requests
that, but you can change it to Japanese in the settings page if you
really want to.

Added the text "under development" to the Japanese option so users
know it's still buggy.